+++ homeassistant/components/modbus/const.py
@@ -2,6 +2,7 @@
 
 from enum import StrEnum
 import logging
+from typing import TYPE_CHECKING
 
 from homeassistant.const import (
     CONF_ADDRESS,
@@ -12,6 +13,10 @@ from homeassistant.const import (
     CONF_SWITCHES,
     Platform,
 )
+from homeassistant.util.hass_dict import HassKey
+
+if TYPE_CHECKING:
+    from .modbus import ModbusHub
 
 # configuration names
 CONF_BAUDRATE = "baudrate"
@@ -162,6 +167,8 @@ DEFAULT_HVAC_ON_VALUE = 1
 DEFAULT_HVAC_OFF_VALUE = 0
 MODBUS_DOMAIN = "modbus"
 DOMAIN = "modbus"
+
+DATA_MODBUS_HUBS: HassKey[dict[str, ModbusHub]] = HassKey(DOMAIN)
 
 ACTIVE_SCAN_INTERVAL = 2  # limit to force an extra update
